requirements.txt
streamlit
pandas
numpy
plotly-express
scikit-learn
joblib
matplotlib
seaborn
streamlit_reveal_slides